•	This system will increase the voting percentage in India. If high security is applied then it may reduce false vote
“ONLINE VOTING SYSTEM” is an online voting technique. It is based on the
other online services like “ONLINE RESERVATION SYSTEM” .In this system
people who have citizenship of INDIA and whose age is above 18 years of
any sex can give his\her vote online without going to any polling booth.
There is a DATABASE which is maintained by the ELECTION COMMISION OF
INDIA in which all the names of voter with complete information is
stored.